Clean run with docker cleanup
